refactor(dart): move widgets into AppRoot
