+++ lib/app_root.dart
@@ -1,23 +1,29 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:brunstadtv_app/graphql/client.dart';
 import 'package:brunstadtv_app/graphql/queries/me.graphql.dart';
+import 'package:brunstadtv_app/theme/bccm_theme.dart';
 import 'package:brunstadtv_app/theme/bccm_typography.dart';
 import 'package:brunstadtv_app/helpers/navigation_utils.dart';
 import 'package:brunstadtv_app/providers/analytics.dart';
 import 'package:brunstadtv_app/providers/settings.dart';
 import 'package:brunstadtv_app/router/router.gr.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:graphql_flutter/graphql_flutter.dart';
+import 'package:on_screen_ruler/on_screen_ruler.dart';
 
+import 'l10n/app_localizations.dart';
 import 'models/auth_state.dart';
 import 'providers/auth_state/auth_state.dart';
 
 class AppRoot extends ConsumerStatefulWidget {
-  const AppRoot({super.key, required this.child, required this.navigatorKey});
+  const AppRoot({super.key, required this.navigatorKey, required this.appRouter, required this.routerDelegate});
 
-  final Widget? child;
   final GlobalKey<NavigatorState> navigatorKey;
+  final AppRouter appRouter;
+  final AutoRouterDelegate routerDelegate;
 
   @override
   ConsumerState<ConsumerStatefulWidget> createState() => _AppRootState();
@@ -116,6 +122,28 @@ class _AppRootState extends ConsumerState<AppRoot> {
 
   @override
   Widget build(BuildContext context) {
-    return widget.child ?? const SizedBox.shrink();
+    return GraphQLProvider(
+      client: ValueNotifier(ref.watch(gqlClientProvider)),
+      child: MaterialApp.router(
+          localizationsDelegates: S.localizationsDelegates,
+          localeResolutionCallback: (locale, supportedLocales) {
+            return locale?.languageCode == 'no' ? const Locale('nb') : locale;
+          },
+          supportedLocales: S.supportedLocales,
+          locale: ref.watch(settingsProvider).appLanguage,
+          theme: ThemeData(),
+          darkTheme: BccmTheme.getThemeData(),
+          themeMode: ThemeMode.dark,
+          debugShowCheckedModeBanner: false,
+          title: 'BCC Media',
+          routerDelegate: widget.routerDelegate,
+          routeInformationParser: widget.appRouter.defaultRouteParser(includePrefixMatches: true),
+          builder: (BuildContext context, Widget? child) {
+            return MediaQuery(
+              data: MediaQuery.of(context).copyWith(textScaleFactor: 1.0),
+              child: !kDebugMode ? child! : OnScreenRulerWidget(gridColor: Colors.white60.withOpacity(0.1), child: child!),
+            );
+          }),
+    );
   }
 }
